+++ NEWS.md
@@ -3,6 +3,7 @@
 * `var_label()`, applied to a data.frame, now accepts a character 
   vector of same length as the number of columns.
 * `set_variable_labels` has a new `.labels` argument.
+* New `unclass` option in `to_factor()`, to be used when `strict = TRUE` (#36)
 
 # labelled 2.0.2